// src/generated-client/models/SignupPostRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Coffee Addiction API
 * Coffee Addiction API swagger documentation
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SignupPostRequest
 */
export interface SignupPostRequest {
    /**
     * Desired username
     * @type {string}
     * @memberof SignupPostRequest
     */
    username?: string;
    /**
     * Password of the user
     * @type {string}
     * @memberof SignupPostRequest
     */
    password?: string;
    /**
     * First name of the user
     * @type {string}
     * @memberof SignupPostRequest
     */
    firstName?: string;
    /**
     * Last name of the user
     * @type {string}
     * @memberof SignupPostRequest
     */
    lastName?: string;
}

/**
 * Check if a given object implements the SignupPostRequest interface.
 */
export function instanceOfSignupPostRequest(value: object): value is SignupPostRequest {
    return true;
}

export function SignupPostRequestFromJSON(json: any): SignupPostRequest {
    return SignupPostRequestFromJSONTyped(json, false);
}

export function SignupPostRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): SignupPostRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'username': json['username'] == null ? undefined : json['username'],
        'password': json['password'] == null ? undefined : json['password'],
        'firstName': json['firstName'] == null ? undefined : json['firstName'],
        'lastName': json['lastName'] == null ? undefined : json['lastName'],
    };
}

export function SignupPostRequestToJSON(json: any): SignupPostRequest {
    return SignupPostRequestToJSONTyped(json, false);
}

export function SignupPostRequestToJSONTyped(value?: SignupPostRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'username': value['username'],
        'password': value['password'],
        'firstName': value['firstName'],
        'lastName': value['lastName'],
    };
}